pyproject.toml
[project]
name = "policy-extractor"
version = "0.1.0"
description = "Extract structured data from insurance policy PDFs using Claude API"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.5",
    "sqlalchemy>=2.0.48",
    "python-dotenv>=1.0.1",
    "pymupdf>=1.27.2",
    "ocrmypdf>=17.3.0",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "loguru>=0.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["policy_extractor*"]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
